fix: re-activation of OProp
In unoptimized React code, the useDerived hook can cause an OProp to be deactivated and reactivated immediately after. Previously, this resulted in the OProp not being added back to its parent OProps.

This commit also removes the logic that deallocated the `watched` map of OProps when it became empty. We might add this back later, but it seems like premature optimization at the cost of complexity.